remove unused props and a bit more refactoring
