--- components/DeleteButton.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const DeleteButton = props => (
-  <form>
-    <input type="submit" value="--" className="button-add" />
-  </form>
-);
-
-export default DeleteButton;
