+++ components/FormCategory.jsx
@@ -1,32 +1,23 @@
 import React, { Component } from 'react';
-import HandleClickInput from './HandleClickInput';
+import HandleClickInput from '../lib/HandleClickInput';
 
-class FormCategory extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {};
-    this.onClickInput = this.onClickInput.bind(this);
-  }
-
-  onClickInput(e) {
-    e.preventDefault();
+const FormCategory = ({ add }) => {
+  const onClickInput = (evt) => {
+    evt.preventDefault();
     const inputElement = document.getElementById('input-cate');
-    HandleClickInput(e, this.props.add, inputElement);
-  }
-
-  render() {
-    return (
-      <form className="form" onSubmit={this.onClickInput}>
-        <input
-          placeholder=" New category"
-          id="input-cate"
-          style={{ display: 'none' }}
-          className="input-gen input-cate"
-        />
-        <input type="submit" value="+" className="button-add" />
-      </form>
-    );
-  }
-}
+    HandleClickInput(evt, add, inputElement);
+  };
+  return (
+    <form className="form" onSubmit={onClickInput}>
+      <input
+        placeholder=" New category"
+        id="input-cate"
+        style={{ display: 'none' }}
+        className="input-gen input-cate"
+      />
+      <input type="submit" value="+" className="button-add" />
+    </form>
+  );
+};
 
 export default FormCategory;
