+++ lib/HandleClickInput.jsx
