--- lib/deleteTodoArr.js
@@ -1,1 +1,0 @@
-function deleteTodoArr(index) {}
